// examples/languages/fsm/src/language-server/generated/ast.ts
export interface Event extends AstNode {
    readonly $container: FSMModel;
    readonly $type: 'Event';
    name: string
}

export const Event = 'Event';

export function isEvent(item: unknown): item is Event {
    return reflection.isInstance(item, Event);
}

export interface FSM extends AstNode {
    readonly $container: FSMModel;
    readonly $type: 'FSM';
    initialState: Reference<State>
    name: string
    states: Array<State>
    transitions: Array<Transition>
}

export const FSM = 'FSM';

export function isFSM(item: unknown): item is FSM {
    return reflection.isInstance(item, FSM);
}

export interface FSMModel extends AstNode {
    readonly $type: 'FSMModel';
    events: Array<Event>
    fsms: Array<FSM>
}

export const FSMModel = 'FSMModel';

export function isFSMModel(item: unknown): item is FSMModel {
    return reflection.isInstance(item, FSMModel);
}

export interface State extends AstNode {
    readonly $container: FSM;
    readonly $type: 'State';
    name: string
    outTransitions: Array<Reference<Transition>>
}

export const State = 'State';

export function isState(item: unknown): item is State {
    return reflection.isInstance(item, State);
}

export interface Transition extends AstNode {
    readonly $container: FSM;
    readonly $type: 'Transition';
    guardEvent?: Reference<Event>
    name: string
    sentEvent: Reference<Event>
    source: Reference<State>
    target: Reference<State>
    time?: number
}

export const Transition = 'Transition';